MASTER-1801, MASTER-1802 refactor, bump version
Refactor!

Formerly, we required the caller to use the with/as
construct to clean up temp files. Now it does this
internally with a simple try/catch.

The external interface is just isign.resign() now. It
may raise isign.NotSignable if you ask it to sign something
that isn't an iOS native app.

Doesn't create any temp files until it actually starts working
on them. Examines .zip and .ipa files with `zipfile` to see if
they are iOS native apps, without unpacking them.

A .zip or .ipa now is considered to *have* an app inside it, rather
than being a subtype of an app. This separates directory structure
and archive format from the business of re-signing.