Make entitlements from existing entitlements
With Watchkit, we need to make sure the namespaces match up
between both app and Watchkit sub-app. This is the first step:
don't blow away the existing ones, just modify them to use
our team id.

Also, the temporary entitlements file was always a lame hack.
There was no need to write to disk, we were just imitating
the usual procedure that one follows with Apple tools. Also,
we just made every app have the same entitlements and that
was bound to fail eventually.