Fix tests for XCode 7+ new codesign output
Tests use Apple's `codesign` to check the contents of a re-signed app.

Previously these tests assumed that
  `codesign -d -r- --verbose=20 SomeApp`
produced lines with unique keys. As of XCode 7+ it seems now they don't;
in particular 'Hash' appears twice.

So we append values to an array and then grep for some values that we want.